requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
flatbuffers==24.3.25
gast==0.6.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.67.1
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
keras==2.15.0
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
ml-dtypes==0.2.0
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pillow==11.0.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.15.0
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.21.0
